requirements.txt
streamlit>=1.15.0
reportlab>=3.6.0
pyyaml>=6.0